Change of the maze display: The walls of the maze is now displayed using buttons instead of lines, which means the walls are now clickable.
Change of the class (architecture):
Delete the MazeGen class and grid class as it makes more sense to put all the maze production related thing in one class and the display related thing in another class.
The new Maze class is a class to keep all the maze related information, like the dimension of it, the entry and exit location, state of walls of each cell and etc...
The MazeWithImage class is a child class extends the Maze class. This class is another type of maze, i.e. a maze with images at entry and exit.
The MazeGeneration class is a class to keep all the methods related to maze generation, optimum solution calculation and so on.
The displayMaze class stores all the method related to the visualization of the maze in the edit tab.

There are several methods in these classes which is empty and going to be finished in the next stage.